Fixed service invocation construction error on iOS devices.
Overwritting the memory for a variable containing the address of an object
appears to leave the object in tact on OS X, it appears to affect the object
in some way on iOS, and causes EXC_BAD_ACCESS when continuing to iterate
through the containing array. This is easily resolved by writing the memory
to a new buffer, instead of being lazy and trying to use the same buffer.